Upgrade Hibernate 3.3.2.GA to 3.5.6-Final.
Do upgrade changes as per Hibernate's 3.5 upgrade guide. These changes
have not been tested.

pom.xml:
- Upgrade Hibernate dependencies from 3.3.2.GA to 3.5.6-Final.

Permission.java, PermissionTemplate.java, Profile.java, ProfileOverrides.java,
Site.java, TargetInstance.java:
- Change type='text' to type='materialized_clob' for SQL types clob (Oracle) or
  text (MySQL and Postgres), as this is the type that now maps to SQL clob.

HarvestCoordinatorImpl.java, InTrayManager.java, InTrayManagerImpl.java,
ProfileManager.java, TargetInstanceManager.java,
TargetInstanceManagerImpl.java, SiteManager.java, SiteManagerImpl.java,
TargetManager.java, TargetManagerImpl.java, InTrayDAO.java, ProfileDAO.java,
ProfileDAOImpl.java, SiteDAO.java, SiteDAOImpl.java, TargetDAO.java,
TargetDAOImpl.java, TargetInstanceDAO.java, TargetInstanceDAOImpl.java,
HomeController.java, SitePermissionHandler.java, TransferSeedsController.java,
MockInTrayDAO.java, MockProfileDAO.java, MockSiteDAO.java, MockTargetDAO.java,
MockTargetInstanceDAO.java:
- Change all methods dealing with Projections.rowCount() from int to long, as
  the underlying HQL type has changed. This means method signatures have
  changed from int to long, Object casts are from Integer to Long, and
  calculations and mocking are now long.

+++ src/main/java/org/webcurator/ui/site/controller/TransferSeedsController.java
@@ -149,7 +149,7 @@ public class TransferSeedsController extends AbstractCommandController {
 		Pagination results = targetManager.searchPermissions(aCriteria);
 		updateCommand(aCommand, aCriteria);
 
-		int numberOfSeeds = siteManager.countLinkedSeeds(aCommand.getFromPermissionOid());
+		long numberOfSeeds = siteManager.countLinkedSeeds(aCommand.getFromPermissionOid());
 
 		ModelAndView mav = new ModelAndView("permission-search");
 		mav.addObject("page", results);
